Stop updating issue record recursively by default
GitHub: fix GH-81

Reported by usk0517. Thanks!!!

It may cause an unique constraint error. If an journal is added to an
issue, the following jobs are enqueued:

  1. UpsertTargetJob(IssueMapper)
  2. UpsertTargetJob(JournalMapper)

If both of them are performed at the same time,
JournalMapper#upsert_fts_target is called twice for the same journal
ID. It causes an unique constraint error with the following steps:

  1. Job A calls find_fts_target that returns a new FtsTarget
  2. Job B calls find_fts_target that returns a new FtsTarget
  3. Job A calls fts_target.save that creates a new record
  4. Job B calls fts_target.save that also creates a new record
     (an unique constraint error)