Add content to issue_contents for for multi-field AND searches (#165)
GitHub: GH-163

The `issue_contents` table was previously used for a "similar issue"
search feature.
But that feature was rarely used. We will change the purpose of this
table for a new "multi-field AND" search across fields.
To support this, we introduce a singular "content" column that has data
from multiple fields for each issue.

We've removed the `contents` column and `is_private` column.
- `contents` is no longer needed because we are unifying data into the
single `content` column.
- `is_private` is removed to simplify the design for now. Private issues
will not be included in the search at all.
- If we think of the private issues, it will be too complicated. So we
will add it again when we really need it.

At the following PRs, we will implement the search logic and add
relevant tests using this table.