Update the status of Journal to the latest
Problem:
When `tag_ids` of an issue is updated, `tag_ids` of Journal is not updated.
The `tag_ids` stores status and tracker information.

Solution:
When `tag_ids` of an issue is updated, the information of Journal is also updated.

Because `options[:recursive]` is not used, the condition is instead `fts_target.saved_change_to_tag_ids`.
We have an issue that has been resolved with https://github.com/clear-code/redmine_full_text_search/commit/2507b826f4a64ba68bb3b0de0d250c46f4163646, so when we update all Journals, we do not create fts_targets.