Update the status of Journal to the latest (#178)
GitHub fixes GH-171

Problem:
When `tag_ids` of an issue is updated, `tag_ids` of journal is not
updated. The `tag_ids` stores status and tracker information.

Solution:
When `tag_ids` of an issue is updated, the information of journal is
also updated.

We have an issue that has been resolved with
https://github.com/clear-code/redmine_full_text_search/commit/2507b826f4a64ba68bb3b0de0d250c46f4163646,
so when we update all journals, we do not create new FTS target.

---------

Co-authored-by: Sutou Kouhei <kou@clear-code.com>
Co-authored-by: Sutou Kouhei <kou@cozmixng.org>